Switch ifBO freeze-thaw step budget from epoch count to wall-clock time
Each thaw step now trains its candidate for at most ifbo_thaw_step
seconds (resuming from its checkpoint), instead of always advancing a
fixed number of epochs. Training is requested up to max_budget every
call so the LR schedule horizon stays fixed, and how far a candidate
actually got is read back from the returned epoch history rather than
assumed. Threads a per-call max_time_seconds override through
train_single_configuration for this.

+++ automl/cli/types.py
@@ -44,4 +44,5 @@ class RuntimeConfig(TypedDict):
     ifbo_greedy_candidate_selection: bool
     ifbo_incumbent_ensemble_top_k: int
     ifbo_incumbent_ensemble_accuracy_threshold: float
-    ifbo_thaw_step: int
+    # Wall-clock cap (seconds) on a single ifBO freeze-thaw step.
+    ifbo_thaw_step: float
